Add all Asian countries in the list
A few of them were there but mostly at random (remnants of the Korean
DC?). Complete the list and point them to eqiad. This doesn't alter
traffic patterns at all, it just makes it a bit more verbose, so we can
make informed decisions on ulsfo and not leave anything out.

Change-Id: I9cef7eb490e01d2df0e8498850cc8763f5a84b18