Switch to the GeoIP City databases & add US states
Now that the GeoIP City databases are available, switch to using them
instead. The freely-available/lite IPv6 database is poor and essentially
has no city/region information, but fortunately gdnsd allows us to
overlay the accurate paid-for IPv4 database.

Add all of the US states (plus D.C.) to the map but pointed to eqiad for
now. This prepares the config for the eventual switch of some states to
ulsfo.

Change-Id: Id4228019667321074b0e21b11431d5cd25f3a401