Geolocate our own IP space manually
Our ARIN IP space is geolocated to San Francisco, which means that with
ulsfo up internal requests get directed to the ulsfo caches. This is
silly, especially for bots from Labs hitting the site, but it also
matters from a couple of other places, such as e.g. the netmapper cron
fetching the Wikipedia Zero list from meta.

Change-Id: I59436ac4ab88ad3d82c47e1edbf7d847206888c3