Switch MX for donate-lb.esams to mchenry/sodium
donate.wikimedia is a GeoCNAME to the various donate-lbs. There was an
inconsistency between MX records for pmtpa/eqiad/ulsfo & esams, with
only the latter pointing directly to aluminium.

Make this consistent across the various lbs, and per discussion with
Jeff, use the mchenry/sodium mail routers instead of aluminium for this.

Change-Id: I06847012f32ae6a5186f3a9f8f287b360c2e537f