Switch uploads to temporary map for ulsfo turnup
uploads traffic that was normally directed to ulsfo
is currently being sent to esams, and needs to be brought
back in chunks.  This is the first step in that process:
define a temporary map based on generic-map which has
all ulsfo (other than WMF OIT) replaced with esams,
use it for the upload-lb decision, and use the correct
ulsfo CNAME there.

Change-Id: I13b39d1142b1ccb720c8e4835d4407874bf1dbe9