Add 'udplog' service alias for fluorine.eqiad.wmnet
UDP logging for the MediaWiki stack is not going away anytime soon, and has to
be configured in multiple places.

  CommonSettings.php:       $wmfUdp2logDest = '10.64.0.21:8420';
  wmerrors.ini:             wmerrors.log_file=udp://fluorine.eqiad.wmnet:8420
  40-mediawiki-apache.conf: & @fluorine.eqiad.wmnet:8420;MediaWiki

Having a service alias will help keep our Puppet code concise (I plan on
adding this service alias for Labs, too, so we'll be able to eliminate a lot of
"if production / elsif labs"-type code), and it will make it easier to move
this service off of fluorine, if it ever needs to be decomissioned.

Change-Id: I33d3e4a58aaea8f83dda4a7bfbe755673aaf1d0d