Remove revdns for unused project-lb.site hostnames
forward was removed in I539244c1 , and these entries
were commented out in I1526dd53 .

note some of these addresses are still actively
configured in puppet for LVS and/or protoproxy.
Work is ongoing there, but I need to make forward
progress on fixing up the zonefiles at this time.

Change-Id: I390e316b16edf4b6f3a017b60e04550c7551c558