Fix GWT text records having explicit zone names
These two zonefiles are symlinked to many other zonefile names
in which the explicit zone name is invalid on the left-hand-side,
so I'm switching them to {{ zonename }} to fix the error.

There's a larger issue to investigate here that this wasn't caught
by our checks or deployment earlier on.  It was noticed while
doing a fresh start of the dns server.

Change-Id: I69194ee64b69fe08825b49e9a78e1a2ff27e6751