Allocate sandbox vlans for codfw and ulsfo
These are not setup yet. The codfw subnet currently overlaps with
the Tampa sandbox vlan, which will be gone in a month. Do not use
for another few weeks, but feel free to allocate IPs in there.

The corresponding IPv6 subnets have been allocated into a new /56 per
site for security reasons, :200::/56. eqiad's only sandbox subnet
doesn't follow that rule, but for consistency, eqiad is getting the new
allocations (for all rows) as well, with an addendum in a comment that
sandbox1-b-eqiad should be renumbered.

Change-Id: Idc8373c5d6eafcd118820606ab154a8859661018