Move mail.wikipedia.org to the main cluster
mail.wikipedia.org right now goes to an ex-Tampa host; move it to the
main cluster where it will be redirected to lists.wikimedia.org.

This removes MX records for it altogether. This is a legacy domain and
according to mail logs, there is no email traffic targetted there.

Bug: 44731
RT:  6981
Change-Id: I152ca6447dccfb7542d2f5bbc8522d179d570a4b