s/eqiad/ulsfo/ as backup for AS/* + RU with esams primary
Of those with esams as the primary, the ones in Asia and RU are
switched from [esams, eqiad] to [esams, ulsfo] as a first-order
approximation of a semi-intelligent way to split traffic with
esams unavailable.

Will use this as a diff baseline for several changes going forward
today for the esams temporary outage and adjust as appropriate.

Change-Id: I879c3bf6389b34f23b21641fa8729efab2d01670