wikitech: make it a CNAME for silver
Rather than using an A record and setting the v4 IP here,
make wikitech a CNAME for the host it is running on, silver,
which now has an A and an AAAA record, so that wikitech.wikimedia.org
becomes IPv6 enabled.

You can compare this to host "radium" and "tor-eqiad-1" for example.

Bug:T73218
Change-Id: If69a47a6de51c888f8582d3dd8da5c6fe9f472c0