disable geolang zerodot for all but wp.org
There are virtually zero hits for $lang.zero outside of
wikipedia.org, but we're still templating it into other
domains like wikibooks, wiktionary, etc.  I don't think this is
really intentional, just an artifact of the shared template.  We
don't have certificates for $lang.zero anywhere but wikipedia, and
AFAIK we've never used or advertised the name for other domains
either.

Change-Id: I8e1a120906f611ffcb64001c38eed5f1e3b84ccb