Switch config to the City variant of GeoIP2
This switches from a combination of the original (version 1) GeoIP City
database, overlaid by a crippled GeoLite variant for IPv6, to the newer,
officially supported GeoIP2 version of the City database, which comes
with (hopefully) proper IPv6 support.

This is in a new format (MMDB) supported only by gdnsd 2.2+ and needs a
new configuration directive, geoip2_db.

Change-Id: I6c447bd0515c4e60269d54c991e1cb3cf8f4c538