Add codfw everywhere on the map
This adds codfw as a backup (mostly to eqiad) everywhere on the map. In
the few cases where we had directed traffic to codfw already (Texas,
Mexico, etc.), it keeps it at the front of eqiad.

The intent of this commit is to give us the ability to set eqiad's
admin_state as "down".

Note that this does not handle trickier situations yet, like the failure
of all of ulsfo, esams and codfw; the map was limited like that before,
in particularly in Asia.

Change-Id: I0274b1bf61d0203b525e18a353d32af823a6e194