Switch Middle-East's backup from ulsfo to eqiad
I very much doubt that esams -> ulsfo (for tier 2) makes much sense.
Most of these countries are geographically closer to eqiad than ulsfo
and I'd expect this to be even more closer in network topology terms. I
have evidence about certain countries in that list (e.g. Cyprus!) but
I'm fairly sure the same holds true for all of them.

This leaves only one remaining oddball with an esams->ulsfo fallback
(and without eqiad on the list at all) in the whole map: the Russian
Federation. Since Russia is huge and we still treat it like a single
geography entry in this map, this probably makes sense. It's hard to
tell exactly as it's likely very dependent on population density etc.

Change-Id: Ic697cd92625a22894212b4b5354c9b35f96678b4