Switch Russia's esams backup from ulsfo to eqiad
Extensive testing with a number of different IPs and different Russian
ISPs revealed that the vast majority of the traffic passes via Europe,
even for the farther ends of Russia like Vladivostok.

Even after engineering the traffic to pass through very specific ISPs
that we know that own submarine cables to Asia (namely: Rostelecom) and
succeeding into passing traffic via Tokyo, we were not able to see a
substantial improvement over eqiad and still a disadvantage over esams.

If we factor in the differences in population density and internet
populations across Russia, it's clear that shifting traffic from esams
to ulsfo rather than eqiad would actually harm performance in times of
esams maintenances.

It's thus clear at this point that even a regional split of Russia with
different GeoDNS per their subdivisions wouldn't help.

Therefore, set Russia to the standard triplet esams,eqiad,codfw like the
rest of Europe.

Change-Id: Ibc588eb3313d4e41091e6f35f72a4f69a54c9b2d