deactivate wikimediacommons.[co.uk|eu|info|jp.net|mobi|net|org]
Unlink all of these from the cluster and don't send traffic anymore,
unless we think it's worth to buy SSL certs for them or you think
the .org is more important than the others.

If you know people who are active on commons and also on Gerrit,
please add them.

Change-Id: I1a33c035111a68c83c7ecd4b4a9c0ee3f1f569fe