Add grafana-admin.wikimedia.org
Grafana 2's native LDAP capabilities require group memberships to be exposed
via some property of the user object. It is not yet capable of searching the
user in specific graphs. <https://github.com/grafana/grafana/pull/2746> fixes
that but it is not yet in master.

So, the plan is to serve Grafana via two vhosts: grafana-admin and grafana.
The former will authenticate visitors via mod_authnz_ldap; the latter would
provide a read-only view to all visitors by setting X-WEBAUTH-USER to some
predefined anonymous read-only account.

Change-Id: I899ef65232c72a7e88a2ee4fc7aabde7a1a4de55