ldap-mirror: Add per DC CNAMEs
Allow per DC CNAMEs, allowing both low latency connections for DC-local
queries and the possibility of an automatic fallback to a secondary, out
of DC, mirror if the client library supports that. Most client libraries
do support that. Add ldap-corp mirrors as well to allow migration to a
slightly better named scheme

Change-Id: I1aa3b95568e622d41643144a3b2c91d9e23ad639