config-geo: ulsfo fallback to codfw before eqiad
This patch is literally just:

s/ulsfo, eqiad, codfw/ulsfo, codfw, eqiad/

As a reasonable approximation without better data, it's probably
best to assume that codfw is a slightly better choice than eqiad
for clients that default to ulsfo, if ulsfo is down and both
primaries are up.  This also gives us better load spread when
ulsfo is down in general.

Bug: T114659
Change-Id: Ie5b0649c1db9deae47f47dd69f3d6be91ff08f88