Kill wmftest.org DYNA, remove text-addrs-v4
There doesn't seem to be a good reason by wmftest.org is pointed to
text-addrs-v4 instead of text-addrs, or why even it's pointed to our
text cluster entirely (the domain is not configured there). This seems
to be just a legacy leftover. Remove this without much fuss -- the
domain has "test" in it anyway :)

Change-Id: Ia78818261b6f5a6629bc40b36dad4e8bdcb2927b