Add "fake" codfw entries for ocg to hotfix an issue
Since ocg signals to mediawiki which server hosts the pdf it should
download via a string like "ocg1001:8000", this was failing on the codfw
appservers as they only have codfw as a search domain. This hack seemed
like a safer hotfix than adding a new search domain on all appservers.

Bug: T133136
Change-Id: I1aa7281a08ef4316e01e0fd1585cf7e787645573