config-geo: default AS=>ulsfo
Similar to the EU situation in ab6a7777 , but I'm a little less
sure about this one.  eqiad's probably not the right choice, but
either of esams or ulsfo could be more-optimal for a variety of
reasons.  Note 'AS' also has a virtual country (in maxmind terms)
called 'AP' for Asia-Pacific, which is a sort of sub-fallback for
that area, which is already mapped to ulsfo as well.

Change-Id: Ia9e6c4904e212263f73c5805947c1c6b1ed193a7