Revert "rhodium: add IPv6 AAAA and reverse"
This reverts commit fb0c6d34cd6f2e4d67912ed0da544b3a3756fd5e.

This caused lag issues with puppet-merge as the SSH connection from
palladium to strontium is not allowed by ferm on IPv6. The ferm rule
present in role::puppetmaster::backend to allow SSH depends on the
puppetmaster having an IPv6 forward DNS record and palladium does not,
hence the issue. We should first try to add an IPv6 forward DNS record
and then revert this revert

Change-Id: I2d5b54a7b94114424fd4e0765e1ed8cb077b9253