add phab1001.eqiad as CNAME for iridium.eqiad
Phabricator is becoming a clustered service.
The new machine in codfw is already called 'phab2001'.

Add a CNAME 'phab1001' for iridium now and start using these
names in firewalling, for reasons as suggested by jcrespo in
the comments on I65fa4f67e2ef39f.

Later when iridium gets reinstalled, rename the host to phab1001
proper, not needing a CNAME anymore.

Change-Id: I250e8b1b0a13aaabad2011d7b242dd456ae5a78c