add git-ssh.codfw.wikimedia.org service IP
We will need git-ssh.codfw.wikimedia.org. equivalent
to the existing git-ssh.eqiad.wikimedia.org. to fix
the puppetization of phab2001.

See hieradata/role/eqiad/phabricator/main.yaml
for the realserver_ips and listen_addresses. These currently
don't exist for codfw and are hardcoded to the eqiad addreses
in modules/role/manifests/phabricator/main.pp.

To make that flexible we have to add this first.

Bug: T143363
Change-Id: I8a439c3e70d90685f9028f18a0b0768d82555a5f