remove gallium.wikimedia.org, keep gallium.mgmt
gallium, the former Continuous Integration server,
has been decom'ed completely and was replaced by
contint1001.wikimedia.org.

It has been shutdown and as the (almost) last step
in the process we are removing the IP from DNS now.

"almost" because there is still the mgmt IP and
datacenter-ops work coming up now to wipe it and
physically remove it from the rack, update racktables etc.

Bug: T95757
Change-Id: Ia14cfdf0448ae3ff7f1aabfec096c84beff8dcae