fix mgmt names in wrong data center
These were more names showing up in the "rejects" log
when running the getmgmtips script, because they existed
in reverse lookup but not in forward.

The first 2 were just missing, but the others were in the
wrong section, so they are in eqiad but ended up with "mgmt.codfw"
names. This should move them into "mgmt.eqiad" where they are
per racktables.

Bug:T149875
Change-Id: I462247e3668a6d681044645b9faf752e9e3e95bb