consistent capitalization of mgmt asset tag names
Sometimes they are "wmf" and other times they are "WMF",
but not in a consistent way. (also forward vs reverse)

When i searched for them with grep i kept missing them or
i had to remember to use grep -i to ignore the case.

So i just searched/replaced all WMF with wmf in my editor.

Change-Id: Ie2125bfa92c17d372adfb88fa3c5dca82eca1338