Add dns alias for analytics and web requests labsdb service
Due to the lack of proper redundancy, this, at this time, is the
easiest and quickest way to provide this service.
Obviously, the proxies themselves are not redundant, but
we do not have the hardware now for that (in fact, we do not have
hardware for these proxies, either, these are borrowed temporarilly
from production).

The good news is that haproxy is quite stable, and the main issue
about ha is not service reliability due to host failure, but the
inability quickly depool databases for maintenance. This would
fix those immediate problems, and starting next quarter we can
maybe do a hardware request and a proper HA implementation of the
proxies with either virtual floating IPs, LVS or dns round-robin.

Bug: T141097
Change-Id: Idb1ccf81cf148a1f1213198510cb26e67bc4c4cb