delete wikitech.m.wikimedia.org
The ticket T120527 is about how the existing
URL https://wikitech.m.wikimedia.org is broken
and shows the portal instead.

To actually make it a working mobile version of
Wikitech it would need T87633 (install MobileFrontend)
but the comments on both say that it can't support
it and/or we don't need it and suggests to delete
the domain.

So here we are deleting it.

Bug: T120527
Change-Id: I803f3ddb34ad1a8ee0abd5e2b5573813e8fd7935