introduce dbmonitor, add dbmonitor[12]001, v4 and v6
Adds IP addresses for <s>tendril</s>,
a generic db monitoring tool,
which is a pre-requisite to creating ganeti instances for it.

As the request ticket says they need to be external IPs,
so we are in wikimedia.org.

Since 2 VMs are requested, one for each DC, it's creating
them as 1001/2001 rather than using element/star hostnames.

Add IPv6 added just because it should be a default for
new things unless there is a reason not to support it.

Bug: T149340
Bug: T149557
Change-Id: I75cbe1e25855b300426fa26aa5a576d0c24163dc