add europium.eqiad.wmnet
I found this orphaned server in racktables and when
i checked mgmt it was booting and it's running an OS
even though it's not in puppet and DNS only had mgmt left.

There was a ticket at T82239 that said back in 2013(!) that
it can be unallocated and reclaimed.

I was looking for a candidate to test installs with for installserver
changes, so i am re-adding this for now and will put jessie on it.

I reopened T82239.

Change-Id: I2b4b73480859bc47547482d8e7f327f51b2cce92