(re-)add hooft as bast3002
(re-)add the server formerly known as hooft,
which had this IP and was removed in Ibef2296b6338c9b58,
as bast3002 to replace bast3001 which has disk issues.

After we get 3002 up and running we can shutdown current 3001
and consider adding a CNAME for convenience.

Rename the remaining "hooft" mgmt interface.

Bug: T131560
Change-Id: I0c5999cc179a81d5a8b3e38844ad092af1da74b0