fix lists/fermium: step 1, remove wrong PTRs for service IP
Remove the wrong PTR remnant for old v4 and v6 service IPs.

These are just a remnant from the past and be removed
separately from other steps.

For v4 it is NOT the .4 IP it is also not bound to an interface
on fermium.

lists.wikimedia.org has address 208.80.154.75
lists.wikimedia.org has IPv6 address 2620:0:861:3::2

75.154.80.208.in-addr.arpa domain name pointer lists.wikimedia.org.

[fermium:~] $ ip a s | grep 208
    inet 208.80.154.74/26 brd 208.80.154.127 scope global eth0
    inet 208.80.154.75/32 scope global eth0
    inet6 2620:0:861:3:208:80:154:74/64 scope global

though:

host 208.80.154.4
4.154.80.208.in-addr.arpa domain name pointer lists.wikimedia.org.

And for IPv6 there are both:

2620:0:861:1::2 and 2620:0:861:3::2

but only inet6 2620:0:861:3::2 is bound to an interface.

Change-Id: Ie63c3a82f3fad2385a84f0fef41bde3b96f6cd1b