add service IP to be used by gerrit slave/standby
A gerrit host has a server IP and a service IP, the latter
being profile::gerrit::server::ipv4 and profile::gerrit::server::ipv6
in Hiera.

cobalt uses 208.80.154.85 (from hiera/role/eqiad), which is gerrit.wikimedia.org
but gerrit2001 uses 208.80.153.106 (from hiera/role/codfw), which is NOT a
service IP but the server IP itself (gerrit2001).

So add a service IP to be used by the standy/slave gerrit server instead.

Change-Id: I6710347ed28fcaa01521a1335b2201d8d6d3d4e9