add "issue globalsign.com" to CAA recs
SSLMate still doesn't list them in the official generator, but
their generator can be downloaded and tweaked to add it (as I've
done here).

The CA-level cutoff date for support is coming up in ~5 weeks, so
GS has to be rolling this out soon.

I have yet to see anything customer-facing from GlobalSign about
their CAA support.  I suspect they perhaps think it's a non-issue
and assume customers who don't know about it will simply not use
CAA (which is a valid choice, and equivalent to allowing all
registrars), but haven't thought about the need to explicitly
communicate the use of the tag "globalsign.com" and get in the
SSLMate generator, etc?

Globalsign does mention using "globalsign.com" as a CAA field in
their recent blog post:

https://www.globalsign.com/en/blog/what-is-the-ca-browser-forum/

... but only as an example of how to configure CAA, not as
documentation on their practices.  So, going with that for now on
the assumption they wouldn't give self-invalid examples.

Bug: T155806
Change-Id: I1fdb793128017f64ea403d3c2cfeade0ded829ea