add releases-jenkins to misc-web cluster
During T164030 we had issues with the Jenkins web UI setup
since releases.wm.org is already active/active in misc-web Varnish,
so it uses geo-DNS.

Result was some people got eqiad and the initial
admin setup while others got codfw with releases2001 and broken
redirects.

We want to keep the existing active/active setup with rsync for
actual release files but need an active/passive setup for Jenkins.

So we are introducting a separate domain for releases-jenkins.

Bug: T164030
Change-Id: I39287d6c3994fdfebbd84c215a9874fbd289fe8e