temp use bast-test.eqiad.wmnet for wmf4727
We still want to test if your 3rd remaining misc
host in eqiad also shows the disk issues that
the other 2 have.

I can't get a DHCP lease in .wikimedia.org due to
switch port config afaict but i can't change that,
so i'm temp using an IP in 10.64.48.0/22 for testing
and then revert it.

Bug: T186623
Change-Id: Id1394c1604b10ea0891fa7d2416e246cf924a6a9