eqsin: BN, BT, CC, CX, KH, KR, LA, MN, MO, MV, TW
These 11 are the last of the "easy" set of specific countries
within the (MaxMind) continent of Asia, where "easy" means routing
to eqsin looks reasonably-close to optimal, eqsin wins over our
other edges on latency, the country is in the initially-defined
set we're attempting turn-up for, and is not blocked by Zero.

Planned for deploy ~ Mon Apr 2nd.

Bug: T189252
Change-Id: I00406145bdd189bd99a2e6fa40129bf70b4a5d18