eqsin: default for AS continent + AP fake-country
The "AP" entry is a pretty clear case, as this is the fake country
used when MaxMind knows the IP is somewhere in the "Asia-Pacific"
sub-region of Asia (in other words, the ~south-eastern portion of
Asia), but doesn't know the specific country, in which case eqsin
is the most-reasonable default at this point.

The default for "AS" at the continent level (when no better data
is known/guessed) is tricky given that parts of Asia (the Maxmind
continent) route better to esams than either of eqsin or ulsfo.
In cases where the IP happens to be in the "AP"-ish area eqsin is
the right answer, while if some of these poorly-located IPs
actually end up way over on the west side (e.g. Jordan) would be
better off with "esams".  The winning arguments for switching this
default to eqsin-first are:

1) The parts of Asia closer to eqsin have the higher net
population (so given no perfect answer, this is the highest net
benefit overall).
2) The existing default before this commit was ulsfo.  Even for
western-asia countries for which esams is the best choice, eqsin
will probably be a better choice than ulsfo was.

Bug: T189252
Change-Id: Id1ae1750f295dd606dd54afdb88110b592743add