Set m4-master to db1107 rather than dbproxy
The eventlogging database replication scheme
(not handled by mysql but by a custom script) forced
the Analytics team to set db1107 (m4-master) replica
(db1108) to be able to accept writes, rather than
staying read only. This caused a weird split brain
scenario in T188991, that we'd like to avoid in the
future. The Eventlogging mysql daemons responsible
to push data to mysql have two nice properties:
1) they read data from kafka, so in case of failures
   we can replay data.
2) data insertion shouldn't lead to duplicates due
   to unique keys present in the table schemas.

This changes the m4-master domain to point directly
to db1107 rather than the current proxy.

Bug: T188991
Change-Id: I59abaa0368a32737c121fed9b1debdf6fedffcc6