add SPF record to disallow email for all parked domains
As requested on T193408, if a domain isn't actually used for
email there should be an explicit SPF record disabling it.

This "parking" template is the zone file that all parked domains
link to.

That's most of the ones listed on https://wikitech.wikimedia.org/wiki/Domains

This change will affect them all at once and we will declare that a parked
domain can/should neve have email.

Bug: T193408
Change-Id: I8fc7feebc0ff8706b5bda407dedc633e5e2f0244