TXT entries for Github domain verification
Github offers a way to verify the URL mentioned in the organization
profile as well as user email address. That is done by adding a couple
TXT entries with a challenge which Github will then inspect to confirm
we own the domain.

Add two TXT entries for each of wikimedia.org (to validate emails) and
www.wikimedia.org (URL on the organization profile).

Bug: T207364
Change-Id: Idd8dfa0b81e04e43087e638907e1be80da986701