Introduce zoterov2 LVS IPs
Add zoterov2.svc.{codfw,eqiad}.wmnet DNS RRs. Those are temporary in
order to allow the citoid service to migrate to the new version of the
zotero translation server while still allowing for a quick rollback.
Later on the v2 part will be dropped and the citoid service will be
moved back to the zotero.svc.{codfw,eqiad}.wmnet RRs

Bug: T201611
Change-Id: If250241344b470dfbdac174cf7ac02e5e1c011ba