add maintenance.eqiad CNAME, point to mwmaint1002
We have "deployment.eqiad.wmnet" that is used and
currently an alias for deploy1001.eqiad.wmnet.

That way there don't have to be hardcoded host names
in other places including user's local SSH config.

But we don't have the same for the maintenance server
and people have to know it is currently mwmaint1002
and will be counting up in the future.

Add maintenance.eqiad.wmnet next to deployment.eqiad.wmnet
and people.eqiad.wmnet.

We probably should not use CNAMEs at all but i also want to be
consistent with the existing setup.

Change-Id: Ib5c962bd6b2164d6d69385ecbba759cc6a3d9a76