Name stray camera IPs as "cam-unknown"
There seem to be PTR records for 10.64.0.4 and 2620:0:861:101::4 pointed
at cam1-a-b-eqiad, but that record's A/AAAA seem to be pointed at
10.64.0.29. Unfortunately, these records go back many years, before
there was a git history, and cameras are currently offline so it's hard
to know if these are stray records or assigned to a device that is
temporarily offline.

Instead of removing them, and potentially reassigning these records to
some other host at a later point and thus risking a collision, name
these temporarily as "cam-unknown". Hopefully this situation will
resolve when T207965 does.

Change-Id: Id35ab131f91ed40568337df41e1b05e923e2f41b