New zone generator gen-zones.py
gen-zones.py differs from authdns-gen-zones.py in several ways:

* Ported to python3
* New dependency on gitpython (python3-git)
* Still uses jinja2 (python3-jinja2)
* Must run from the repo root
* Single argument for output zones directory
* Output zones dir must be empty of contents if defined
* If output zones dir is not set, dry-run mode
* Unconditionally processes everything into fresh empty dir
* {{ serial }} timestamps are deterministic based on git history.
  Specifically, they're generated from the committer datestamp of
  the last commit to touch the real template, the zone's link to
  the real template, and/or any included sub-templates.  They also
  include the git short hash and commit title in a DNS comment

Because this generates output only to a fresh empty directory, the
serials are deterministic from git, we gain several benefits:

* With full regeneration on every run, we avoid the previous
  pitfalls that lead to issues with e.g. adding languages to
  templates resulting in no update to the affected zones.
* With deterministic serials from git, we avoid the previous
  pitfalls with new/reimaged servers and other situations where
  serial numbers could go out of sync between authdnses
* With the added commentary reference to the git log entry, we
  avoid (at least for local observers, but not over the protocol)
  ambiguity about multiple changes in the 1 hour resolution window
  available to easily-human-readable serial numbers, and also
  ensure that every applicable commit causes a change in the zone
  file's contents.

The check-gdnsd.sh script invoked for CI and/or local testing is
switched to use the new script immediately, but the puppet-driven
authdns-local-update process is not (yet).

authdns-gen-zones.py is left in place for now to ease the
remaining transition process.  From here we'll fix up the
authdns-local-update system to use this script to generate into an
empty temporary directory and then use local rsync to force the
updates over to /etc/gdnsd/ and track whether changes happened to
the actual zones and/or config (as opposed to scripts and
metadata) so that it can take appropriate action (currently it
replaces if config files were updated, and otherwise
unconditionally reloads zones even if unecessary).

Change-Id: Ie5834f3403b4c802e9232823c9e7d928004d85f3