deploy-check.py replaces check-gdnsd.sh
This replaces what check-gdnsd.sh was doing with a python3 script
which also can do real deployments with the --deploy flag, and
thus also can replace the bottom portion of authdns-local-update
from puppet.

As compared to authdns-local-update, the deployment done here is
much more careful and thorough with a 2-stage design that re-uses
the same code for staging to a temporary directory that the
checker uses, but copying in the real deployed puppet config
from the host itself rather than the mocked versions used for CI.

authdns-local-update always overwrote the system's copies first,
then tried to half-undo things if the checkconf failed.  This
script can also be used for direct local deployments from the git
clone in cases where ssh, dns, and/or authdns-update are broken.

Change-Id: Ibabe8e06a5d01f911a4a8db85366006846dbe136