Add analytics dbstore SRV records
This is a proposal to ease the use of the new dbstore100[3-5]
hosts from a user perspective. As described in
https://wikitech.wikimedia.org/wiki/Analytics/Data_access#MariaDB_replicas
we are moving from analytics-store.eqiad.wmnet to a multi-ip/port
environment, and some automation will need to be written in several
places to avoid cumbersome settings or procedures. We already have
CNAMEs but it would be great if there was more friendly way to get
what port needs to connect to the right mysql instance, hence I
thought about using the SRV records. Not sure if this is a misuse of
this record type, asking for feedback to the broader team.

Bug: T212386
Change-Id: I8dfb148fa45854535d302fc473a8950e1164db5c