ns[012].wikimedia.org: 1D TTLs
Upstream registrar serves our NS records and the matching glue address
records with 1D TTLs, and we should be doing the same.  We have 1D on
the NS records, but not the addresses.

This value is better anyways as it increases resilience when a
recursor has problems reaching us, and NS IPs don't change that
quickly and easily anyways (believing our own 1H would be fooling
ourselves).  Also updated the matching PTRs to 1D as well.

Change-Id: I684b8c193e9568cdff172a12bee8779681eaf3e4