wm.org: 1h non-dyna records for foo.dcname entries
All of these are better off being explicit addresses.  They're not
actually EDNS-scoped or dynamic, and this makes the forward and
reverse entries actually match up in a simple way.  Also bumped
the 600 TTLs on them to 1H (they don't fail over).

There are other issues in these stanzas (like, why do the
donate-lb variants even exist, and why do they have MX records?),
but this is a step in the right direction at least.

Change-Id: If3fa33f110b41dccbcd71180721f1f02a11d3b3e