wiktionary: test with zone-local CNAME->DYNA
Note this and the referenced langlist-c template from the previous
commit use TTL=300, which is half our normal value.  This should
ease any reversion, and also pushes a little harder on DNS cache
expiries to help uncover real-world issues with the two-query
zone-local-CNAME model.

If there are issues with wiktionary accessibility due to DNS
issues around 2019-04-08 -> 2019-04-15, then this experiment was a
failure and just this DNS commit needs reverting to fix wiktionary
for users.

Bug: T208263
Change-Id: I50ba2f7697b4e7717aaece286259f7576dc04f8d