wikipedia.org: test with zone-local CNAME->DYNA
This is a follow-on to the successful experiment in f1cc2682 , but
with the broader audience of wikipedia.org.  It's otherwise
identical in nature except that all TTLs are set to 600 instead of
300 to help analysis.

What we expect is that cold caches manually following CNAME chains
would roughly double our authserver query rate, but then this will
be offset back in the negative direction by the ~672x reduction in
the count of distinct wikipedia.org wiki hostnames having edns
client subnet fanout in recursor caches, which is hard to predict
the impact of due to all kinds of randomness (popularity curve of
the hostnames, regionality of caches, user population behind
various caches, etc).

I expect a net reduction in the overall, but of unknown magnitude,
and not as much reduction as we would see in a production version
of this experiment with long CNAME TTLs and all projects included.

In case of credible/numerous user reports of DNS resolution
failures for wiki hostnames within wikipedia.org, reverting just
this change should restore everything to a working state for those
users within ~10 minutes.

This experiment should run for at least a week to give a chance to
gather and analyze such failure reports if they're rare, and then
perhaps we should up the CNAME (but not DYNA) TTLs substantially
to measure that impact before we move on to a final production
version of this scheme for all projects.

Bug: T208263
Change-Id: I961b9091cf8244f6317a8b0e0c6057cba6f828c9