wikipedia.org CNAME experiment: 4H CNAMEs
The initial experiment started with the deploy at
~2019-04-18T16:20 of I961b9091cf8244f6317a8b0e0c6057cba6f828c9
with the DYNA TTL remaining at our standard 600 (10 minutes) and
the CNAME TTLs also at 600 for quick reversion if early/widespread
problems were detected.

This change is intended to deploy 24H later (assuming we continue
to observe no faults or user reports of breakage), and raises the
CNAME TTLs to 4H.  This should allow further analysis of the
query-rate impact of this scheme, while still keeping the
reversion time within somewhat-reasonable bounds.  In the final
non-experimental version we're likely to target a full 1D for the
CNAME TTLs.

Bug: T208263
Change-Id: I9dee96d07a74d8b57b09333305b65922f44b2a4a