wmnet: Point m2-master.codfw to dbproxy2002
dbproxy2002 is now ready in codfw m2.
So m2-master.codfw.wmnet should point to it.

mysql --skip-ssl -hdbproxy2002.codfw.wmnet -e "select @@hostname; show global variables like 'read_only'"
+------------+
| @@hostname |
+------------+
| db2044     |
+------------+
+---------------+-------+
| Variable_name | Value |
+---------------+-------+
| read_only     | ON    |
+---------------+-------+

Also db2078:3312 is the slave in m2, so point m2-slave to it.
The client should handle the port and connect to 3322 as that is the
host for db2078 in m2

Bug: T176532
Bug: T202367
Change-Id: Ic4f2ccbb2258235772e140a266e7bf015f61953b