wmnet: Point m1-master to dbproxy1014
dbproxy1001 and dbproxy1006 will need to be decommissioned.
dbproxy1014 is in place, but it is on a rack that will be affected by
the PDU maintenance in October.
I want to move m1-master to dbproxy1014 for a few days to make sure it
works fine, so we can decommission dbproxy1006 (knowing that dbproxy1014
works as expected).
Once dbproxy1014 is known to work fine, we can move back to dbproxy1001
until dbproxy1014 PDU is replaced

Also changing the TTL to make this be applied faster, and also faster to
revert if needed.

Bug: T202367
Change-Id: I94485f3da32a0ef8f530482d80cf428f8345a3f7