wikimediacloud.org: add initial zone file
The wikimediacloud.org is a new domain that will be used for WMCS cloud service
APIs.

The initial use case for this is providing a higher level domain name for
openstack API requests. The FQDN openstack.<region>.wikimediacloud.org will map
to the load balancer running on the primary openstack controller for that
region.

The openstack controllers are not configured with a virtual IP. In the event of
a failure or planned maintenance, this host entry will need to be remapped to
the secondary controller.

Bug: T223907
Change-Id: I7cdba8ed8482f167e7732984b02415560ffd2af7