move gerrit-new service IP to B network
The new Gerrit server is in the B row so
the service IP to go with it also needs to be
in this network and C is wrong.

We are using .136 and .137 in the same network
then for gerrit-new and gerrit1001.

Also move IPv6 reverse record for gerrit1001
into the right section for the B network.

Bug: T222391
Change-Id: If6473f0245b35dc1946f2c3d260634a30e8f2e8f