DNS Cleanup: single line SOA with comment split
This rids our SOAs of the multi-line parenthetical record
continuation and the leading whitespace within it.  The templated
serial stuff is split into two separate template items: the raw
serial number for use in the singular SOA line, and a separate
comment line to put above the SOA in each file.

Change-Id: I3485a904d2600bf98f0df7314c15220b0781b46a