cloud: introduce delegation for codfw1dev.wikimedia.cloud
The domain `codfw1dev.wikimedia.cloud` is going to be use to hold FQDNs for VM
instances running in the codfw1dev openstack deployment.
The authoritative DNS server for that domain should the Designate daemon
running in the codfw1dev openstack region.

Previous to this patch, we didn't have a domain delegation because there wasn't
a proper FQDN for the auth DNS server @ codfw1dev. So this patch introduces
them:

 * ns0.openstack.codfw1dev.wikimediacloud.org
 * ns1.openstack.codfw1dev.wikimediacloud.org

Bug: T243556
Bug: T243766
Change-Id: Ie066736f23246b7da1a8b0abfd3213cd040dbf2d
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>