add irc2001.wikimedia.org
A DNS entry is needed before the VM can be created.

Per ticket comments we are still using a public IP for now.

Even though it's just a single machine introducing a new cluster
name.

Or we can use another star name if you'd prefer that.

The new cluster name would have to be added to:

https://wikitech.wikimedia.org/wiki/Infrastructure_naming_conventions#Servers

Bug: T244719
Change-Id: Id3ec4235c9f97fdbb6e3df93c73da1508a3a5d0e