wikimediacloud.org: introduce new service names for DNS
New service names for DNS running inside openstack (designate).
Only add new names, don't drop old ones until we are ready to do so.

This patch only adds PTR records for the IPv6 addresses that don't have a
previous record. For those having a previous record, it will be updated when we
are ready to drop the old names.

Bug: T243766
Change-Id: Ie7f58fd1f13ea8b7d16a5a5c508ac66ce7a85bee
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>