remove install1003/2003 again to recreate with public IPs
After further discussion we are going back to public IPs
for these new install servers because of the webproxy part.

We might split out webproxies to dedicated machines later but
not for now and they do need to keep public IPs like installservers
had before as well for that.

We have to decom the VMs and recreate them again in wikimedia.org.

Bug: T224576
Bug: T242602
Change-Id: I83f308169441f7a03f5ed34342e69713a1b63518